linear chirp image generation macro
Makes an image with a wave of increasing
spatal frequency.

This can then be gaussian filtered to
show limited resolution of a
frequency band limited system
such as a microscope objective lens.

Sample Output images are included in the .png
screeshot, with gaussian blur of sigma 7
and also then added noise
to show noise effect on practical resolution.