ConvDeconv js demo: refactor test image drawing
factor out the math functions to generate the test images into
javascript functions, our of the function that handles the general image
drawing.

an image definition function defining the shape of the pattern is passed
to the abstracted function that runs the loops to generate the pixels at
all the x,y locations and puts the array into an IJ image.