conv deconv demo js: add the diff image to temp
updat the temp image by adding the difference image
not the blurred scaled version of it.
That was a bug!
Iterations now improve the result image as expected,
and high frequencies are resotored.
Noise is still a problem.... beed to filter it
with a Gaussian smooth every 5 iterations or so.