algebraic update working, plus smoothing every 5 iters
theh slgorithm description calls for smoothing every 5 iterations to prevent build up of noisein the guess image.
We use the same global sigma as in the initial raw image smoothing.
This number is a magic parameter so would need to be adjusted according to image signal:noise perhaps? .
im testing with noise images, but sigma is still only 0.7